requirements.txt
django==4.1.7
django-celery-beat==2.4.0
django-compressor==4.3.1
django-cors-headers==3.13.0
django-cors-middleware==1.5.0
django-extensions==3.2.1
django-filter==22.1
django-tagulous==1.3.3
django-guardian==2.4.0
djangorestframework==3.14.0
django-crispy-forms==2.0

# Other Python and project related libraries
flower==1.2.0
requests==2.28.2
amqp==5.1.1
psycopg2-binary==2.9.5
redis==4.5.1
watchdog==2.2.1
drf-nested-routers==0.93.4
s3fs==2023.1.0
minio==7.1.13
celery==5.2.7
Pillow==9.4.0 --global-option="build_ext" --global-option="--disable-tiff" --global-option="--disable-freetype" --global-option="--disable-lcms" --global-option="--disable-webp" --global-option="--disable-webpmux" --global-option="--disable-imagequant" --global-option="--disable-xcb" --global-option="--disable-zlib"
flatten-json==0.1.13
PyYAML==6.0
Markdown==3.4.1

uvicorn==0.20.0
#For Oauth2 remove comment
#django-oauth-toolkit
#oauthlib
#social-auth-app-django
#social-auth-core
